Backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pypdf==3.17.1
together==0.2.5
cartesia==1.4.0
ffmpeg-python==0.2.0
pydantic==2.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
